+++ .formatter.exs
@@ -1,11 +1,8 @@
+# This file is synced with stordco/common-config-elixir. Any changes will be overwritten.
+
 [
-  inputs: ["*.{ex,exs}", "{lib,test}/**/*.{ex,exs}"],
+  import_deps: [:patch],
+  inputs: ["*.{heex,ex,exs}", "{config,lib,priv,test}/**/*.{heex,ex,exs}"],
   line_length: 120,
-  export: [
-    locals_without_parens: [
-      assert_kafee_message_produced: 0,
-      assert_kafee_message_produced: 1,
-      assert_kafee_message_produced: 2
-    ]
-  ]
+  plugins: []
 ]