+++ lib/kafee/producer.ex
@@ -125,7 +125,7 @@ defmodule Kafee.Producer do
       use Supervisor
 
       @doc false
-      @impl true
+      @impl Supervisor
       # credo:disable-for-lines:15 Credo.Check.Design.AliasUsage
       def init(init_opts \\ []) do
         # credo:disable-for-lines:2 Credo.Check.Warning.UnsafeToAtom
@@ -138,9 +138,15 @@ defmodule Kafee.Producer do
           |> Kafee.Producer.Config.validate!()
 
         children = [
-          {Kafee.Producer.Config, config},
-          {config.producer_backend, config}
+          {Kafee.Producer.Config, config}
         ]
+
+        child_spec = config.producer_backend.child_spec([config])
+
+        children =
+          if is_nil(child_spec),
+            do: children,
+            else: Enum.reverse([child_spec | children])
 
         Supervisor.init(children, strategy: :one_for_one)
       end
@@ -185,8 +191,8 @@ defmodule Kafee.Producer do
 
   ## Examples
 
-      iex> normalize([%Kafee.Producer.Message{key: "test", partition: nil}], MyProducer)
-      [%Kafee.Producer.Message{key: "test", partition: 0, partition_fun: :random}]
+      iex> normalize([%Kafee.Producer.Message{key: "test", partition: nil, topic: "test"}], MyProducer)
+      [%Kafee.Producer.Message{key: "test", partition: 0, partition_fun: :random, topic: "test"}]
 
   """
   @spec normalize([Message.t()], atom()) :: [Message.t()]