+++ lib/kafee/producer/sync_backend.ex
@@ -7,43 +7,23 @@ defmodule Kafee.Producer.SyncBackend do
 
   @behaviour Kafee.Producer.Backend
 
-  use Supervisor
-
   alias Kafee.Producer.Config
 
-  @doc false
-  @impl true
-  def init(%Config{} = config) do
+  @doc """
+  Child specification for the lower level `:brod_client`.
+  """
+  @impl Kafee.Producer.Backend
+  def child_spec([config]) do
     brod_endpoints = Config.brod_endpoints(config)
     brod_client_opts = Config.brod_client_config(config)
 
-    children = [
-      %{
-        id: config.brod_client_id,
-        start: {:brod_client, :start_link, [brod_endpoints, config.brod_client_id, brod_client_opts]}
-      }
-    ]
-
-    Supervisor.init(children, strategy: :one_for_one)
-  end
-
-  def init(opts) do
-    raise ArgumentError,
-      message: """
-      The `Kafee.Producer.SyncBackend` module expects to be given
-      a `Kafee.Producer.Config` struct on startup.
-
-      Received:
-      #{inspect(opts)}
-      """
-  end
-
-  @doc """
-  Starts a new `Kafee.Producer.SyncBackend` process and associated children.
-  """
-  @impl true
-  def start_link(%Config{} = config) do
-    Supervisor.start_link(__MODULE__, config, name: Kafee.Producer.Backend.process_name(config.producer))
+    %{
+      id: config.brod_client_id,
+      start: {:brod_client, :start_link, [brod_endpoints, config.brod_client_id, brod_client_opts]},
+      type: :worker,
+      restart: :permanent,
+      shutdown: 500
+    }
   end
 
   @doc """
@@ -58,7 +38,7 @@ defmodule Kafee.Producer.SyncBackend do
       {:ok, 1}
 
   """
-  @impl true
+  @impl Kafee.Producer.Backend
   def partition(%Config{brod_client_id: brod_client_id}, message) do
     with {:ok, partition_count} <- :brod.get_partitions_count(brod_client_id, message.topic) do
       partition_fun = :brod_utils.make_part_fun(message.partition_fun)
@@ -69,7 +49,7 @@ defmodule Kafee.Producer.SyncBackend do
   @doc """
   Calls the `:brod.produce_sync/5` function.
   """
-  @impl true
+  @impl Kafee.Producer.Backend
   def produce(%Config{} = config, messages) do
     for message <- messages do
       :telemetry.span([:kafee, :produce], %{topic: message.topic, partition: message.partition}, fn ->