+++ test/kafee/producer/async_backend_test.exs
@@ -1,12 +1,16 @@
 defmodule Kafee.Producer.AsyncBackendTest do
   use Kafee.KafkaCase
 
-  setup %{topic: topic} do
-    Application.put_env(:kafee, :producer,
+  defmodule MyProducer do
+    use Kafee.Producer,
       producer_backend: Kafee.Producer.AsyncBackend,
-      topic: topic
-    )
+      partition_fun: :random
+  end
 
+  setup %{topic: topic} do
+    Application.put_env(:kafee, :producer, topic: topic)
+
+    spy(Kafee.Producer.AsyncWorker)
     start_supervised!(MyProducer)
     :ok
   end
@@ -14,7 +18,7 @@ defmodule Kafee.Producer.AsyncBackendTest do
   describe "produce/2" do
     test "sends messages" do
       messages =
-        for num <- 1..10 do
+        for num <- 1..5 do
           %Kafee.Producer.Message{
             key: to_string(num),
             value: to_string(num)
@@ -22,6 +26,7 @@ defmodule Kafee.Producer.AsyncBackendTest do
         end
 
       assert :ok = MyProducer.produce(messages)
+      assert_called(Kafee.Producer.AsyncWorker.queue(_pid, _messages), 1)
     end
   end
 end