+++ test/kafee/producer_integration_test.exs
@@ -1,14 +1,20 @@
 defmodule Kafee.ProducerIntegrationTest do
-  use Kafee.BrodCase
+  use Kafee.KafkaCase
 
   import ExUnit.CaptureLog
 
   # Generally enough time for the worker to do what ever it needs to do.
-  @wait_timeout 1000
+  @wait_timeout 1_000
+
+  defmodule MyProducer do
+    use Kafee.Producer,
+      producer_backend: Kafee.Producer.AsyncBackend,
+      partition_fun: :random
+  end
 
   setup %{topic: topic} do
     start_supervised!(
-      {MyIntegrationProducer,
+      {MyProducer,
        [
          hostname: KafkaApi.host(),
          port: KafkaApi.port(),
@@ -20,18 +26,18 @@ defmodule Kafee.ProducerIntegrationTest do
   end
 
   describe "large messages" do
-    test "logs and continues" do
+    test "logs and continues", %{topic: topic} do
       message_fixture = File.read!("test/support/example/large_message.json")
-      large_message = String.duplicate(message_fixture, 4)
+      large_message = String.duplicate(message_fixture, 10)
 
       log =
         capture_log(fn ->
           assert :ok =
-                   MyIntegrationProducer.produce([
+                   MyProducer.produce([
                      %Kafee.Producer.Message{
                        key: "something_huge_above_4mb",
                        value: large_message,
-                       topic: "wms-service",
+                       topic: topic,
                        partition: 0
                      }
                    ])
@@ -47,11 +53,11 @@ defmodule Kafee.ProducerIntegrationTest do
       log =
         capture_log(fn ->
           assert :ok =
-                   MyIntegrationProducer.produce([
+                   MyProducer.produce([
                      %Kafee.Producer.Message{
                        key: "something_small",
                        value: "aaaa",
-                       topic: "wms-service",
+                       topic: topic,
                        partition: 0
                      }
                    ])